User-to-user trust and cached taxon geoprivacy

+++ app/controllers/users_controller.rb
@@ -7,8 +7,8 @@ class UsersController < ApplicationController
     :unless => lambda { authenticated_with_oauth? },
     :except => [ :index, :show, :new, :create, :activate, :relationships, :search, :update_session ]
   load_only = [ :suspend, :unsuspend, :destroy, :purge,
-    :show, :update, :relationships, :add_role, :remove_role, :set_spammer,
-    :merge ]
+    :show, :update, :followers, :following, :relationships, :add_role,
+    :remove_role, :set_spammer, :merge, :trust, :untrust ]
   before_filter :find_user, :only => load_only
   # we want to load the user for set_spammer but not attempt any spam blocking,
   # because set_spammer may change the user's spammer properties
@@ -282,7 +282,7 @@ class UsersController < ApplicationController
   def show
     @selected_user = @user
     @login = @selected_user.login
-    @followees = @selected_user.friends.where( "users.suspended_at IS NULL" ).paginate(:page => 1, :per_page => 15).order("id desc")
+    @followees = @selected_user.followees.page( 1 ).per_page( 15 ).order( "users.id desc" )
     @favorites_list = @selected_user.lists.find_by_title("Favorites")
     @favorites_list ||= @selected_user.lists.find_by_title(t(:favorites))
     if @favorites_list
@@ -314,14 +314,27 @@ class UsersController < ApplicationController
       format.json { render :json => @selected_user.to_json( opts ) }
     end
   end
+
+  def followers
+    @users = @user.followers
+    @users = @users.page( params[:page] ).order( :login )
+    counts_for_users
+    respond_to do |format|
+      format.html { render :friendship_users, layout: "bootstrap" }
+    end
+  end
+
+  def following
+    @users = @user.followees
+    @users = @users.page( params[:page] ).order( :login )
+    counts_for_users
+    respond_to do |format|
+      format.html { render :friendship_users, layout: "bootstrap" }
+    end
+  end
   
   def relationships
-    @users = if params[:following]
-      @user.friends.where( "users.suspended_at IS NULL" ).paginate(page: params[:page] || 1).order(:login)
-    else
-      @user.followers.where( "users.suspended_at IS NULL" ).paginate(page: params[:page] || 1).order(:login)
-    end
-    counts_for_users
+    @friendships = current_user.friendships.page( params[:page] )
   end
   
   def get_nearby_taxa_obs_counts search_params
@@ -776,16 +789,40 @@ class UsersController < ApplicationController
     redirect_back_or_default( root_path )
   end
 
+  def trust
+    if friendship = current_user.friendships.where( friend_id: params[:id] ).first
+      friendship.update_attributes( trust: true )
+    else
+      friendship = current_user.friendships.create!( friend: @user, trust: true, following: false )
+    end
+    respond_to do |format|
+      format.json { render json: { friendship: friendship } }
+    end
+  end
+
+  def untrust
+    if friendship = current_user.friendships.where( friend_id: params[:id] ).first
+      friendship.update_attributes( trust: false )
+    end
+    respond_to do |format|
+      format.json { render json: { friendship: friendship } }
+    end
+  end
+
 protected
 
   def add_friend
     error_msg, notice_msg = [nil, nil]
     friend_user = User.find_by_id(params[:friend_id])
-    if friend_user.blank? || friendship = current_user.friendships.find_by_friend_id(friend_user.id)
+    if friend_user.blank? || friendship = current_user.friendships.where( friend_id: friend_user.id, following: true ).first
       error_msg = t(:either_that_user_doesnt_exist_or)
     else
       notice_msg = t(:you_are_now_following_x, :friend_user => friend_user.login)
-      friendship = current_user.friendships.create(:friend => friend_user)
+      if friendship = current_user.friendships.where( friend_id: friend_user.id ).first
+        friendship.update_attributes( following: true )
+      else
+        friendship = current_user.friendships.create( friend: friend_user, following: true )
+      end
     end
     respond_to do |format|
       format.html do
@@ -801,7 +838,11 @@ protected
     error_msg, notice_msg = [nil, nil]
     if friendship = current_user.friendships.find_by_friend_id(params[:remove_friend_id])
       notice_msg = t(:you_are_no_longer_following_x, :friend => friendship.friend.login)
-      friendship.destroy
+      if friendship.trust?
+        friendship.update_attributes( following: false )
+      else
+        friendship.destroy
+      end
     else
       error_msg = t(:you_arent_following_that_person)
     end
